01-unpaid-bills-merge/pyproject.toml
[tool.poetry]
name = "01-unpaid-bills-merge"
version = "0.1.0"
description = ""
authors = ["Manish Sinha <arithmetic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
sendgrid = "^6.11.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
